+++ CHANGELOG.md
@@ -9,8 +9,9 @@ All notable changes to this project will be documented in this file.
 - Base class for exceptions raised by this library
 - Documentation checks using ``ruff``
 - Exception: ``ProblemDefinitionError``
+- Plan: ``first_input`` getter
+- Plan: ``is_empty`` property
 - Problem: setter for the initial state
-- Solution: ``first_input`` getter
 
 ### Changed